Moved splitter to age_splitter and chnaged subsequent calls in test
